requirements/base.txt
# Base requirements
Unidecode==0.04.17
pytz==2014.10
python-dateutil==2.4.0
requests==2.5.1
Django==1.7.4
django-allauth==0.19.1
django-simple-history==1.5.4
django-aggregate-if==0.5
django-adminplus==0.3
django-cron==0.3.6
django-multiselectfield==0.1.3
django-widget-tweaks==1.4.1
django-compressor==1.5
django-libsass==0.4
django-bootstrap-sass==0.0.6-alpha
jsonfield==1.0.2
openpyxl==2.2.6

# Dependencies for base requirements
six==1.9.0                    # python-dateutil, django-aggregate-if, django-libsass
python-openid==2.2.5          # django-allauth
requests-oauthlib==0.4.2      # django-allauth
oauthlib==0.7.2               # django-allauth
South==1.0.2                  # django-cron
django-common-helpers==0.6.4  # django-cron
django-appconf==1.0.1         # django-compressor
libsass==0.8.3                # django-libsass
jdcal==1.2                    # openpyxl
python-magic==0.4.15          # python-magic

# Dependencies between base requirements
# requests==2.5.1             # django-allauth
# django-compressor==1.5      # django-libsass

# Installed by 'virtualenv env' directly
# argparse==1.2.1
# wsgiref==0.1.2
